requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
codetiming==1.3.0
Flask==2.0.2
frozenlist==1.2.0
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Levenshtein==0.16.0
lyricsgenius==3.0.1
MarkupSafe==2.0.1
multidict==5.2.0
polling==0.3.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.2
python-Levenshtein-wheels==0.13.2
rapidfuzz==1.8.2
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.3
spotipy==2.19.0
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
